+++ src/PluginProcessor.cpp
@@ -198,7 +198,7 @@ void AudioPluginAudioProcessor::handleNoteOn( MidiKeyboardState *pSource, int mi
       m_Voices.erase( midiNoteNumber );
    }
 */
-   int vel = 127 * velocity;
+   int vel = (int)( 127 * velocity );
 
    std::list<Sample *> s = getSamplesByMidiNoteAndVelocity( (size_t)( midiChannel - 1 ), midiNoteNumber, vel );
    for( Sample *pSample : s )
@@ -370,7 +370,7 @@ void AudioPluginAudioProcessor::processBlock( juce::AudioBuffer<float>& buffer,
          float *pLeft = buf.getWritePointer( 0 );
          float *pRight = buf.getWritePointer( 1 );
 
-         if( !pVoice->process( pLeft, pRight, buf.getNumSamples(), m_sampleRate ) )
+         if( !pVoice->process( pLeft, pRight, buf.getNumSamples(), (float)m_sampleRate ) )
          {
             stoppedVoices.insert( pVoice );
          }
